Update controller tests

+++ src/main/java/es/jmjg/experiments/application/user/DeleteUser.java
@@ -3,7 +3,9 @@ package es.jmjg.experiments.application.user;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import es.jmjg.experiments.application.shared.exception.Forbidden;
 import es.jmjg.experiments.application.user.dto.DeleteUserDto;
+import es.jmjg.experiments.application.user.exception.UserNotFound;
 import es.jmjg.experiments.domain.repository.UserRepository;
 
 @Service
@@ -16,7 +18,17 @@ public class DeleteUser {
   }
 
   @Transactional
-  public void deleteByUuid(DeleteUserDto deleteUserDto) {
+  public void delete(DeleteUserDto deleteUserDto) {
+    if (!deleteUserDto.userDetails().isAdmin()) {
+      throw new Forbidden("Only admin users can delete users");
+    }
+
+    var user = userRepository.findByUuid(deleteUserDto.uuid());
+
+    if (user.isEmpty()) {
+      throw new UserNotFound(deleteUserDto.uuid());
+    }
+
     userRepository.deleteByUuid(deleteUserDto.uuid());
   }
 }
